Initital port of mkssi-fast-export sources
Generally things are working, with the exception of binary files which
are currently ignored and not exported.

There is a known bug related to revision history comments and daylight
savings time.

Reading RCS revisions is not being done very efficiently, which should
be addressed.

Some sort of author map needs to be added.

In general, the code could use clean-up.