Optimizations and bug fixes
The major optimization here is in how the file revision data is
exported.  Previously, the data was exported inline with the commit
which included that revision of the file.  That was an inefficient
approach to large files with many revisions.  For example, if foo.txt
has two thousand revisions, it takes a while to work backward to rev.
1.1.  If one commit includes foo.txt rev. 1.1, we were working our way
back from the head revision to rev. 1.1; then if the next commit was
foo.txt rev. 1.2, we started over and worked our way back from the head
revision to rev. 1.2.

The new method is to export all file revision data as blobs up front,
and have the commits refer back to them.  So for our foo.txt example, we
work our way back from the head revision to rev. 1.1 only once, and then
never need any of its revision data again.

In support of this, rcs.c provides a new interface.  Instead of a
function which implements "fetch this revision of this file", it has a
function that implements "fetch all revision of this file".

rcs.c was also altered so that its rcs_line structure no longer has to
malloc a copy of every line; this was observed to be slow.  Instead,
each line points into the original string buffer, until it needs to be
modified.  This also means the lines are now terminated by a newline or
a NUL, instead of always a NUL.

Miscellaneous bug fixes:

export.c was not traversing the project branch list correctly.  It
handled branches which were branched from the trunk, but not branches
which were branched from another branch.

import.c was letting some of the weird MKSSI vc_XXXX.000 files slip
through (it filtered out those in the project root, but not in
lower-level directories), which was causing failures because some of
those files have RCS metadata errors.

Fix daylight savings time bug in lex_date(): let the system detect
whether DST applies, rather than instructing that it does not.