Add parameter to allow for trunk history on branch
There exists an MKSSI project where the trunk revision history, at a
given point, was switched from a trunk project revision to a nameless
branch project revision.  For example, imagine that the trunk history is
in project.pj revisions 1.1 through 1.886, but branches at that point,
such that it is on 1.886.1.1 - 1.886.1.n.

I do not know how this project got into that state.

Nor do I understand how MKSSI knows to find the trunk history on that
branch.  Nothing points at this project.pj branch revision which
contains the trunk history; it is a nameless branch.

To deal with this one project, I have added a parameter to specify a
project.pj branch revision to use as a continuation of the trunk
history.  This is ugly, since it requires that the user examines the
project.pj and manually find this branch number, something which
requires knowledge of MKSSI internals.