Minor enhancements, code clean-up, commenting
A couple changes worth noting...

export.c:
Enhance detection of executable files.  In addition to file extension,
the file contents are also examined for a shebang or ELF magic number.

merge.c:
Greatly simplify code which generates commit messages, mostly via
utility functions.  No longer try to keep the first line under 80
characters in auto-generated commit messages; such long lines are
possible with long file paths, but it does not happen often enough to
warrant the extra checks or the inconsistent commit messages.  Remove
the non-operative code to merge changes based on their MKSSI label; in
practice, when the label is the same, so is the check-in comment.