Fix to expand the $Log$ keyword
Revision history was being added after the $Log$ keyword, but the
keyword itself was not being expanded.  On most files this made no
difference since the source already contained the $Log$ keyword in its
expanded form, but a few files had a bare $Log$ or a $Log$ keyword with
the wrong file name.