Fix bugs with changeset creation and merging
Fix a theoretical (not observed) issue in adjust_deletes() where the
timestamp check was backward, so we might have failed to record an
update to a file that was subsequently deleted in the same checkpoint.

Fix the check in merge_updates() to ensure that a file is not updated
more than once in the same commit.  There was code for that already
present, but it was overly simplistic.

Also update merge_updates() to ensure that a later revision to a file is
never committed prior to an earlier revision.