Fix issues with newlines
1) Most MKSSI RCS files use Unix newlines for everything, but a handful
of files have been observed that have Windows newlines in the patch
text.  MKSSI automatically converts everything to the preferred newline
setting, but Git was sticking with the original newlines, at least with
how my Git is configured.  Update to implicitly convert these Windows
newlines to Unix newlines so that everything is consistent.

2) Files that did not end in a newline were having a newline implicitly
appended.  Add code to prevent this, to preserve the original file
contents more faithfully.