Relax restrictions on name case fixups
There are sometimes discrepencies between the case used in the file
system names and the case used in the MKSSI project.pj file.  In every
observed instance, names which are UPPERCASE on the file system are
lowercase in the project.  This was being corrected, but only for file
names (not directories) and only if the file name looked 8.3.  Relax
these restrictions and trust that the name case used in the project is
accurate for everything.