Copy buggy MKSSI log behavior for '@' characters
In the log (i.e., the revision history comment), as elsewhere, any '@'
character must be escaped to "@@", because RCS uses the '@' character as
a delimiter.

When expanding the $Log$ keyword and copying the log text, the "@@"
sequences should be changed back to a single '@' character, but MKSSI
fails to do so in this context.  In other contexts, such as when
displaying the log text from the GUI, it does happen.

This program must copy this buggy behavior in order to exactly duplicate
the MKSSI project.