Reinstate a kluge to fixup file name case
The "name" member of struct rcs_file is used when expanding RCS keywords
(such as $Log$ and $Source$), and as such needs to have the correct
capitalization.

This fix does not work correctly if the file name capitalization changes
over time.  If that comes up, then some way of finding the canonical
name during keyword expansion will need to be devised.