Move code to avoid escaping @s in log pre/postfix
'@' characters in the log text need to be escaped (doubled) in order to
mimic buggy MKSSI behavior; see commit 63dbccf0.

However, this was being done in the wrong place, such that '@'
characters on the prefix/postfix parts of the line were also being
doubled.  This caused a problem in a Windows batch file where the $Log$
keyword appeared on a line like below:

@REM $Log$

The "@REM " is the prefix that gets copied into the inserted lines.  The
'@' character is not doubled by MKSSI, but this program was doubling it.
Avoid this by calling rcs_data_reescape_ats() prior to adding the prefix
and postfix text to the log lines.