Add support for "f" member types
Unlike "a" member types, which explicitly list the file revision, "f"
member types seem to always retrieve the head revision.  This was
determined by observation and might not be correct in all cases, since
"f" member types are rare and thus the sample size was small.

Previously "f" member types were being ignored, but this resulted in
missing files in the Git project, compared to the MKSSI project.