Better warnings for encrypted files
Previously we were reporting encrypted archives as "corrupt", which is
not really true.  However, this program does not support these encrypted
archives.  Even if we wanted to add support, that might be tricky since
the encryption algorithm used by MKSSI is not documented in the manual,
and there is the possibility that it is proprietary.  However, it is
unlikely that I will attempt to support encrypted archives; around here,
only very old projects are using this feature, and I doubt anyone
remembers the passwords anyway.

This commit makes the program display a more accurate warning message
for encrypted archives.  It is a warning message instead of an error
message, since converting the rest of the project is potentially still
useful.