Fix $Log$ for duplicate of duplicate revisions
When the $Log$ RCS keyword is expanded for a duplicate revision, the log
text of the revision from which the duplicate revision was duplicated
must be included.  This was implemented in commit 906504ee.

That commit, however, incorrectly assumed that if the revision was a
duplicate of a duplicate revision, that the log text of the original
non-duplicate revision should also be inserted.

For example, it assumed that for duplicate rev. 1.2.1.1.1.1, that after
inserting the log text for duplicate rev. 1.2.1.1, it should also insert
the log text for rev 1.2.

MKSSI does not behave that way.  For duplicate rev. 1.2.1.1.1.1, it will
insert the log text for duplicate rev. 1.2.1.1, but it stops there.
Arguably, MKSSI _should_ also insert the log text for rev. 1.2, but it
does not, and we aim to be bug-for-bug compatible with MKSSI.