Fix daylight saving time edge case
Previously we were storing file revision timestamps in Unix time (a
time_t value).  When we needed a date/time string for RCS keyword
expansion, we converted the time_t into a string.  This almost always
worked.  However, there was an edge case where timestamps that fell into
the hour skipped by moving the clock forward for DST were converted into
a string that was an hour off from the original.  This was causing
content differences between the original MKSSI project and the converted
Git repository.

To fix this, the date/time string is now generated directly from the
date/time fields in the RCS metadata; and RCS keyword expansion can use
these stored strings.  This requires more memory, but not much, compared
to how much is used already.