Change behavior for text-mode "f" member types
Commit 1f563bca added support for the rare, poorly-documented "other"
project member type, denoted by "f" in project.pj.

The assumption, based on what I had observed up to that point, was that
MKSSI always grabbed the head revision for "other" members.  This was a
false assumption.

The sample size I am working with is very small: eleven "other"-type
members from only two projects.  It is difficult to determine the
correct behavior of "other" project members from such a small sample,
but nonetheless...

Of the eleven "other"-type members, ten are binary files.  For each of
these, grabbing the head revision of the file produces the correct
behavior (in some cases, the head revision is also rev. 1.1; but not in
every case).  The eleventh is a text file, and the correct behavior is
to grab rev. 1.1 -- but without RCS keyword expansion.  (Note that
binary files never have RCS keyword expansion.)

So the current implementation does exactly that: grabs the head revision
for binary files, and rev. 1.1 without keyword expansion for text files.
This may not be correct in all cases, especially for text files, since
the sample size is one.

Exporting a text revision without RCS keyword expansion is something we
never had to do before, so this required some new and somewhat ackward
code to be written.