Fix capitalization adjustment bug
If a project had a file and directory such a these:

  dir/foo/
  dir/FOOBAR.txt

The file dir/FOOBAR.txt was being outputted as dir/fooBAR.txt.

The problem was that the capitalization-adjustment code assumed that the
length in struct dir_path included the path separator, which was indeed
what was intended; however, due to an off-by-one error in the function
dir_list_from_path(), the path separator was excluded from the length.