Export tip revisions
By parsing the MKSSI project directory, in addition to the MKSSI RCS directory,
it is possible to export the _complete_ history for the MKSSI project, including
changes which have never been captured in an MKSSI checkpoint.  Thus, it is no
longer necessary to checkpoint every MKSSI branch prior to exporting its
revision history with mkssi-fast-export.

Also, fixed a couple of weird MKSSI branch cases: 1) branches can have the same
revision number, and the changes for those revisions need to be exported to all
branches sharing that revision number; and 2) MKSSI tolerates broken branches
which exist in MKSSI metadata but not in RCS metadata.