Use MKSSI timestamps for adds/updates
The revision timestamp that MKSSI stores for an added or updated file is the
file's mtime, which can be considerably older than when the user actually
checked-in the add or update.  Previously, mkssi-fast-export ignored these
timestamps and used the timestamp of the subsequent checkpoint.  While this
resulted in sequential commit timestamps, the timestamps could be grievously
incorrect if the checkpoint was much newer than the file revisions.  It was also
a problem when exporting tip revisions (see commit 744e1a7bfe), since there is
no subsequent checkpoint -- we were just using time() for those changes.

Instead, use the MKSSI timestamp for commits that add or update files.  When
multiple file revisions are merged into the same add/update commit, the newest
timestamp is used for the commit.  Although this can result in non-sequential
commit timestamps, it seems like the better trade-off.

MKSSI does not store timestamps for file deletions, so those commits are still
using the timestamp of the next checkpoint (or time() if there isn't a next
checkpoint).