Fix bug with tip history and --trunk-branch
If --trunk-branch was specified, the tip (uncheckpointed) history for branches
older than the trunk branch was not being exported correctly.  Instead, we
exported a bogus commit which deleted every file on the branch.

The problem was that the for loop in project_read_tip_revisions() was using
project_branches before it had been populated with anything other than the
master branch.  Thus, any branches inserted into the project_branches list
(which is sorted by revision number) prior to the master branch were ignored.
Since the per-branch list of tip file revisions has a NULL default value, we
generated a commit to delete all the files on those ignored branches.

Fix this problem by parsing the tip revision of the master branch before looping
through project_branches.

Also, refactor to reduce code duplication.