Export tip revisions only with --proj-dir
Normally --proj-dir should be specified, but it might be necessary to omit it if
the project directory is missing or corrupt.  If --proj-dir is omitted, then the
tip revisions are unknown and can't be exported.  We were attempting to export
those revisions anyway, which results in a commit that deletes every file.