Better timestamp for tip deletions
MKSSI has no timestamps for file deletions.  Normally, mkssi-fast-export uses
the timestamp of the next checkpoint as the timestamp for the deletion commit.
But what about changes from the tip of the branch, where there is no subsequent
checkpoint?  Previously, we were using the current time (the return value of
time()), but it's better to use the mtime of the project file for that branch,
which will usually correspond to the last time that branch was modified, and
which will be closer in time to when those files were actually deleted.