Tolerate missing RCS patches
If a corrupt RCS file has one or more missing patches, instead of aborting with
a fatal error (as before), export an empty file.  This emulates what MKSSI does
when the RCS file is corrupt in that manner.

Caveat: This has only been lightly tested, since not many projects have RCS
files with missing patches.