More name capitalization fixes
Commit 0e6e2cd66d implemented implicit renames for directories whose name
capitalization changes over time.  This commit implements that logic for regular
files, since it is needed for at least one project.

In addition, fixed a bug where a deletion within a renamed directory used the
wrong path.

On a cosmetic level, changed the author of rename commits from "Unknown" to
"mkssi-fast-export", since it is the tool which is generating these commits.