Fix stale usage
Since commit 744e1a7bfe, the RCS directory has been provided as an option,
rather than as a fixed argument.  Update the "usage:" line accordingly.

Also, if there are unrecognized arguments on the command line, print them out
before printing the usage, so that it's obvious why the usage is being printed.