Don't export demarcating tags for empty branches
If there are no commits on a branch, attempting to tag that branch will cause
git fast-import to throw an error.  To avoid that, only export demarcating tags
to branches that have at least one commit.

Primarily, this avoids spurious errors when mkssi-fast-export is used on a fresh
MKSSI project that has no history.

Also, since we are now counting how many commits are on each branch, print those
commit counts when the export is complete, for informational purposes.