Fix bug with multiple renamed parent dirs
If a renamed file or directory had _multiple_ parent directories that were
renamed in the same commit, the old path wasn't being updated correctly, leading
to a git fast-import fatal error because of a path that didn't resolve.

Also fixed a bug with the parent directory matching: "foo" is a parent directory
of "foo/bar.txt" but _not_ "foobar.txt".  It's necessary to check for the path
separator.