Fix merge of added file with missing RCS patch
commit_msg_adds() will throw an error if an add based on a missing RCS patch is
merged with any other adds.  However, merge_adds() was only partly checking for
missing patches.

By the way, the whole idea of not merging adds or updates based on missing RCS
patches is that, because this is corrupted project history, it should be in its
own commit to make it stand out.