Rework branch creation
Don't export a branchpoint until just before exporting project revisions for
that branch.

For duplicate branches, don't create the branch until all the changes have been
exported to the duplicated branch, rather than independently exporting the same
set of changes to all duplicate branches.  Notably, this eliminates the ugly
checkpoint name munging that we were doing before.

Update the statistics to track the not only the total number of commits on a
branch, but also the number of commits that were original to that branch.