Fix deletion path adjustment
When adjusting a deletion path to account for capitalization renames, make sure
to use the longest matching path.  Also, for directory renames, make sure the
directory path is really a parent directory, not just a path which starts with
the same characters.

Basically the same fixes that were in commit 86d1a5f0b4, applied to the
adjust_deletes_for_renames() function.