Fixed a branch ambiguity bug
Perhaps this can be best explained by example: there is a project which has two
branches, which we'll call "A" and "B".  In the variant block in project.pj, the
revision number for both "A" and "B" is 1.145.1.66.  However, "A" is using
project revisions from the 1.145.1.66.1.x line, whereas "B" is using project
revisions from the 1.145.1.66.2.x line.  This was not being handled correctly,
since both lines were being exported to both branches.  Instead, we need to
export the *.1.x changes to "A" and the *.2.x changes to "B".  To know which
branch is *.1.x and which branch is *.2.x, we need to save the tip revision
number from the branch's vpNNNN.pj file (I'm not aware of this being listed
anywhere else).  Then, we can use those tip revisions numbers to disambiguate
the branches when exporting the project revisions.