New RCS keywords, project.pj file name cleanup
The following new RCS keywords are implemented in this commit:

- $Locker$
- $ProjectName$
- $ProjectRevision$
- $RCSfile$

The two $Project*$ keywords are obviously MKSSI extensions.  $ProjectName$
unfortunately cannot be expanded correctly without an additional parameter to
mkssi-fast-import.

$Locker$ is expanded under the assumption that the file is unlocked.  If this
tool is ever used on a project where that keyword is used in a locked file, then
something would need to be done about that.

$ProjectName$ needed the name of project.pj in the project directory with the
canonical capitalization.  Thus, I took the opportunity to clean a few things
up.  The canonical capitalization of project.pj (in both the RCS directory and
the project directory) and project.vpj (in the project directory) are now saved
early on and used later.  Previously, only "project.pj" and "PROJECT.PJ" were
supported; now other variants like "Project.pj" will also work.  Also previously
the project.pj and project.vpj in the project directory were assumed to be lower
case (or located on a case-insensitive network file system); now they can be
any case.  mkssi-fast-export is now also tolerant of a missing project.vpj
directory.