Fix $Log$ expansion with trailing chars
The line with the $Log$ RCS keyword is the template for leading and trailing
characters for the log lines.  If the $Log$ line had trailing characters, they
weren't always being copied into the log lines correctly: if the $Log$ keyword
expanded to a longer string, the line buffer that we were using as the template
would no longer have the trailing characters in their original positions.  To
fix this, make a copy of the line prior to keyword expansion and use that as the
template.