Fixes for "other" member archives
Previously it was assumed that the "other" member archive, on a binary file,
would check-out the head revision.  However, apparently it actually checks-out
the copy of the file in the project directory.  Normally those would have the
same data, but not in a corrupt project.  In order to mimic the behavior of
MKSSI, we need to export a blob for the project directory copy of the file, and
use that for the "other" member archive of binary files.

Furthermore, files with the "other" member archive can be checked-out even if
the RCS master for the file is missing: MKSSI just grabs the copy in the project
directory.  That behavior is now emulated by this program.  One problem with
such files is that this program uses struct rcs_file for all file references,
which doesn't work well when there isn't an RCS file.  The (somewhat awkward)
work-around is to have "dummy" instances of struct rcs_file for those files
which are being exported from the project directory and don't have an RCS
master.