Support store-by-reference format
From the MKSSI online help:

  Storing (Entire File) by Reference [...]

  In the Reference format, the archive file only contains a summary of each
  revision with a pointer to a file that contains that revision.  [... I]t
  improves performance when checking large binary files in or out of a project.

Files with this format have an extra "storage" field in their RCS header with a
subdirectory name that contains the file revisions.  Each file in that
subdirectory is named after the revision number.  The contents of those files
are the contents of each revision.

This commit implements the store-by-reference format only for binary files.  It
could be implemented for text files too, if there is a need.

Also in this commit, if an RCS file starts with "--MKS-Archive--", skip over
that header before parsing the RCS metadata.  This is related to the store-by-
reference format, since such headers have been observed on store-by-reference
RCS files.