Use sorted list for canonical directory name
Use the first appearance of a directory name in the _sorted_ list of file
revisions as the canonical directory name, rather than using the raw list from
the project data.

With the old method, the directory capitalization in the Git repository
sometimes differed from the directory capitalization in the MKSSI sandbox.  The
new method, in preliminary testing, seems to fix that problem.