Fix RCS keywords that depend on project revision
A given file revision (e.g., foo.c rev. 1.7) can be referenced by many project
revisions -- i.e., by many branches and checkpoints.  Usually, the file revision
data is not affected by which project revision is referencing it.  We take
advantage of that to export blobs for each file revision and reuse those blobs
for all the references to that file revision.

Unfortunately, there are edge cases where that approach doesn't work, because
the RCS keywords will expand differently depending on which project.pj revision
is being used to check-out the file.  The obvious example is $ProjectRevision$,
which expands to the project.pj revision being used for check-out.  For example,
if two checkpoints both reference foo.c rev. 1.7, and that revision has a
$ProjectRevision$ keyword, the file data for foo.c rev. 1.7 won't be the same
for those two checkpoints.

In addition to $ProjectRevision$, the RCS keywords that expand to a file path or
file name -- $Header$, $Id$, $RCSfile$, $Log$, and $Source$ -- can also expand
differently depending on the project revision.  The file path or name always
uses the name capitalization from the project.pj file listing, rather than the
capitalization in the project or RCS directories.  However, different revisions
of project.pj may reference the same file revision with different capitalization
variants for the path or name.  For example, suppose that bar.c rev. 1.3 has an
$RCSfile$ keyword.  If a project revision has bar.c rev. 1.3 listed as "BAR.C a
1.3", then $RCSfile$ will expand to "$RCSfile: BAR.C $".  But if another project
revision has bar.c rev. 1.3 listed as "bar.c a 1.3", then $RCSfile$ will expand
to "$RCSfile: bar.c $".  Thus, bar.c rev. 1.3 can have different contents,
depending on how its listed in the project.

This commit implements handling for these edge cases.  It detects cases where
the contents of a file revision are dependent on the project revision; and, for
those cases, it creates commits that export the contents of the affected file
revisions "just-in-time", in the context of the project revision currently being
exported.

$ProjectRevision$ updates are exported in their own commits (unless the affected
file is already being updated, in which case the $ProjectRevision$ update is
implicit).

Capitalization changes to path/name RCS keywords are bundled with the rename
commits which actualize those capitalization changes.

The traditional method of exporting reusable blobs is retained for file
revisions that are unaffected by these edge cases, since this is much faster
than doing a just-in-time export of everything.