Autodetect --trunk-branch if possible
If the project directory is available, then --trunk-branch can be autodetected
from the revision number in the project.pj file in the project directory.

--trunk-branch is retained as a parameter in case it is needed for a project
whose project directory is lost.

If the project directory is specified, but --trunk-branch is nonetheless
explicitly specified, then the autodetection is skipped, giving precedence to
the explicitly specified --trunk-branch value.