Ignore "other" members which don't exist
Project members with the "other" member type can be exported from the project
directory, even if they don't exist in the RCS directory.  However, if the file
also doesn't exist in the project directory, or if it is not a regular file in
the project directory (e.g., it's a directory name), then the member needs to be
ignored.

This fixes a failure with a project that had an "other" member which pointed at
a directory, leading to a fatal error in file_buffer() when we attempted to
read that file to export its contents.